==> data_and_databases_project/jupyter_notebooks_funakoshi/Part 1 - Paralympic medallists (summer) - final.ipynb <==
# Lede Summer 2019 Project - Part 1
## Make a database of female and male Paralympic all-time medallists (summer)
* get years only, not start_year and end_year

import requests
import pandas as pd
import re
import numpy as np
import os

import itertools

from bs4 import BeautifulSoup
from selenium import webdriver

#### LINKS:
Summer:
https://www.paralympic.org/sdms4/hira/web/multiMedallists/type/summer

Winter:
https://www.paralympic.org/sdms4/hira/web/multiMedallists/type/winter

url = 'https://www.paralympic.org/sdms4/hira/web/multiMedallists/type/summer'
response = requests.get(url)
para_doc = BeautifulSoup(response.text)

### Get names of paralympic athletes

# Get athlete names (full name)
para_info = para_doc.find_all('tr')
full_name = []

for athlete in para_info:
    try:
        full_name.append(athlete.find_all('td')[1].string)
    except:
        pass

# full_name

# Get athlete LAST names
para_info = para_doc.find_all('tr')
last_names = []

for name in full_name:
    last_names.append(re.findall(r'([A-Z]+[ -]?[A-Z]{2,}).*', name)[0])

# last_names

# Get athlete FIRST names
para_info = para_doc.find_all('tr')
first_names = []

for athlete in para_info:
    try:
        first_names.append(athlete.find_all('td')[1].string.split(' ')[-1])
    except:
        pass
    
# first_names

### Get citizenship (country code) of paralympic athletes

# Get citizenships of paralympic athletes

para_info = para_doc.find_all('tr')
para_citizenship = []

for athlete in para_info:
    try:
        citizenship = athlete.find_all('td')[3]
        no_citizenship = len(citizenship.find_all('a'))    
        if no_citizenship == 1:
            para_citizenship.append(citizenship.text)

    ## Check for multinationals
        elif no_citizenship == 2:
            multi = []
            multi.append(citizenship.text[:3])
            multi.append(citizenship.text[3:])
            para_citizenship.append(multi)
        else:
            print('Not single or dual citizenship, or there is some other problem')
    except:
        pass
    
# para_citizenship

### Get the year/s during which the paralympic medallists competed

para_info = para_doc.find_all('tr')

years_dict = {}
years = []
for athlete in para_info:
    try:
        year = athlete.find_all('td')[4].string
        years.append(year)
    except:
        pass
# years

### Get total number of medals for paralympic athletes

para_info = para_doc.find_all('tr')
medals_total = []

for athlete in para_info:
    try:
        medals_total.append(athlete.find_all('td')[8].text)
    except:
        pass

# medals_total

### Gold medals for paralympic athletes 

para_info = para_doc.find_all('tr')
medals_gold = []

for athlete in para_info:
    try:
        medals_gold.append(athlete.find_all('td')[5].text)
    except:
        pass

# medals_gold

### Silver medals for paralympic athletes 

para_info = para_doc.find_all('tr')
medals_silver = []

for athlete in para_info:
    try:
        medals_silver.append(athlete.find_all('td')[6].text)
    except:
        pass

# medals_silver

### Bronze medals for paralympic athletes 

para_info = para_doc.find_all('tr')
medals_bronze = []

for athlete in para_info:
    try:
        medals_bronze.append(athlete.find_all('td')[7].text)
    except:
        pass

# medals_bronze

### Assign gender for para athletes

para_info = para_doc.find_all('tr')
gender = []

for athlete in para_info:
    try:
        gender.append(athlete.find_previous('h2').text.strip())
    except:
        pass

# gender

### Put all para info into one list of dictionaries, then into a dataframe

para_info = para_doc.find_all('tr')
rows = []

for athlete in para_info:
    row = {}
    try:
        full_name = athlete.find_all('td')[1].string.strip()
        row['full_name'] = full_name
        row['last_name'] = re.findall(r'([A-Z]+[ -]?[A-Z]{2,}).*', full_name)[0]
        row['first_name'] = athlete.find_all('td')[1].string.split(' ')[-1]
        
        row['medals_total'] = athlete.find_all('td')[8].text.strip()
        row['medals_gold'] = athlete.find_all('td')[5].text.strip()
        row['medals_silver'] = athlete.find_all('td')[6].text.strip()
        row['medals_bronze'] = athlete.find_all('td')[7].text.strip()
        
        row['gender'] = athlete.find_previous('h2').text.strip()
        row['other_info'] = ''
        row['alternate_name'] = ''
        
        years = athlete.find_all('td')[4].string
        row['years'] = years

        citizenship = athlete.find_all('td')[3]
        no_citizenship = len(citizenship.find_all('a'))    
        if no_citizenship == 1:
            row['citizenship'] = citizenship.text.strip()

    ## Check for multinationals
        elif no_citizenship == 2:
            if 'GER' in citizenship.text.strip():
                row['citizenship'] = citizenship.text[:3]
            else:
                multi = []
                multi.append(citizenship.text[:3])
                multi.append(citizenship.text[3:])
                row['citizenship'] = multi
        else:
            row['citizenship'] = 'Not single or dual citizenship, or there is some other problem'
    except:
        pass
    rows.append(row)

rows

# Remove empty rows
rows = list(filter(None,rows))

df_summer = pd.DataFrame(rows)
df_summer

df_summer.to_csv('para_summer.csv', index=False)

## Manually edit the csv to create duplicate rows for HARRIMAN Margaret (double citizen)
* Make duplicate row for Margaret Harriman -- one with full_name as HARRIMAN Margaret (RHO) and the other as HARRIMAN Margaret (RSA) so that the dataframes merge properly when joining with the df of country codes
* Manually add the comment 'Also competed representing Rhodesia' or 'Also competed representing South Africa' in the other_info column
* Re-import the cleaned csv

df_summer_clean = pd.read_csv('para_summer_cleaned.csv')

df_summer_clean.shape

## The sport event is missing from this website!
* Add events for every athlete in a new text file, making sure to name Harriman's duplicate rows as HARRIMAN Margaret (RHO) or HARRIMAN Margaret (RSA) so dataframes merge properly

event = pd.read_csv('para_summer_events.txt', sep=',')

event.shape

merged = df_summer_clean.merge(event, how='inner', left_on='full_name', right_on='name')
merged = merged.drop(columns = 'name')

merged.shape

merged

merged['game_type'] = 'Paralympic'
merged['season'] = 'Summer'

## Clean full_name for Paralympic athletes so that it's formatted as first_name LAST_NAME


merged['full_name'] = merged.first_name + ' ' + merged.last_name
merged

merged.to_csv('para_summer_complete.csv', index=False)
