==> data_and_databases_project/jupyter_notebooks_funakoshi/Part 4 - Olympic medallists (winter) - final.ipynb <==
# Lede Summer 2019 Project - Part 4
# Make a database of female and male Olympic all-time medallists (summer)
A LOT OF ISSUES WITH THE ORIGINAL FILE. Misspelling, different spellings, different names used, etc.

In the text file, cleaned:
* Missing/wrong data for Clas Thunberg
* Ahn Hyun-Soo (Victor AN) -- changed nationality from S.Korea to Russia.
* Eliminated duplicate row for Marit BJØRGEN and Sven Kramer

import requests
import pandas as pd
import re
import numpy as np
import os

import itertools

from bs4 import BeautifulSoup
from selenium import webdriver

#### LINKS:
https://www.paralympic.org/sdms4/hira/web/multiMedallists/type/summer

SUMMER:
https://stillmed.olympic.org/media/Document%20Library/OlympicOrg/Factsheets-Reference-Documents/Games/Records-and-Medals/Factsheet-Records-of-medals-at-the-Games-of-the-Olympiad.pdf#_ga=2.24637754.1432842169.1561491999-1688621044.1561491999

### Turn pdf to text

# Turn pdf into txt

def pdf_to_text(name):
    folder = "/Users/minami/Documents/Classwork/projects/Final-Project-Lede-Summer2019/my_work/"
    input1 = folder + name
    txt_name = name.replace(".pdf",".txt")
    output1 = folder + txt_name
    os.system("pdftotext '%s' '%s'" % (input1, output1))

pdf_to_text('oly_winter.pdf')

## Manually clean the text file using regex

* Standardise format of rows by adding columns, replacing dashes with commas, eliminating white space, formatting years so that they follow \d\d\d\d-\d\d\d\d format

cols = ['gender', 'medal_count', 'last_name', 'first_name', 'citizenship', 'event', 'years']
df = pd.read_csv('oly_winter_most.txt', sep=',', names=cols, index_col=False)
df.head(3)
df.shape

df

### Fill in values for gender

df['gender'] = df.gender.fillna(method='ffill')
df = df.dropna(subset = ['last_name'])
df

### Make a new column with the full name

df['full_name'] = (df.last_name+df.first_name).str.strip()
# df.full_name

# Scrape medals data for athletes
* Make endpoints for urls
* Then clean up the endpoint for the extra dashes, Jr.s
* Make a list of endpoints over which I can interate
* Clean up the list of endpoints for athletes with a middle name or a hyphenated last names

df['endpoint'] = df.full_name.str.replace(' ','-')

### Make a list of endpoints


athletes = df.endpoint.to_list()

### Clean the list

* Hyun Soo Ahn also goes by Victor An (and also changed citizenship from S.Korea to Russia)
* Special characters are replaced with regular alphabets in the url endpoints
* Some people don't have their full names as their endpoints (in the url) -- Marja-Liisa-KIRVESNIEMI-HAEMAELAEINEN, for example, has an endpoint of marja-liisa-hamalainen

for idx, name in enumerate(athletes):
    if name == 'Hyun-Soo-AHN-(Victor-AN)':
        athletes[idx] = 'Victor-AN'
    elif name == 'Victor-AN-(AHN-Hyun-Soo)':
        athletes[idx] = 'Victor-AN'
    elif name == 'Marit-BJØRGEN':
        athletes[idx] = 'marit-bjoergen'
    elif name == 'Ole-Einar-BJØRNDALEN':
        athletes[idx] = 'Ole-Einar-BJOERNDALEN'
    elif name == 'Andrea-EHRIG-SCHÖNE-MITSCHERLICH':
        athletes[idx] = 'andrea-mitscherlich'
    elif name == 'Gunda-NIEMANN-STIRNEMANN':
        athletes[idx] = 'gunda-kleemann'
        
    elif name == 'Marja-Liisa-KIRVESNIEMI-HAEMAELAEINEN':
        athletes[idx] = 'marja-liisa-hamalainen'
    elif name == 'Elena-VÄLBE':
        athletes[idx] = 'elena-vaelbe'
    elif name == 'Bjørn-DÆHLIE':
        athletes[idx] = 'bjorn-daehlie'
    elif name == 'Emil-SVENDSEN':
        athletes[idx] = 'emil-hegle-svendsen'
    elif name == 'Eero-MÄNTYRANTA':
        athletes[idx] = 'eero-mantyranta'
    elif name == 'Karin-KANIA-ENKE':
        athletes[idx] = 'karin-enke'
    else:
        pass
    
# athletes

### Edit the endpoints in the original dataframe so that they can be merged on endpoints later on!!!

df['endpoint_cleaned'] = athletes

rows = []
for athlete in athletes:
    try:
        url = f'https://www.olympic.org/{athlete}'
        response = requests.get(url)
        doc = BeautifulSoup(response.text)
        medals = doc.find(class_='medal-box').text.strip().split( ' \n\n\n ')
        
        row = {}

        gold = int(medals[0][1:])
        silver = int(medals[1][1:])
        bronze = int(medals[2][1:])
        
        row['endpoint'] = athlete
        row['medals_gold'] = gold
        row['medals_silver'] = silver
        row['medals_bronze'] = bronze
        row['medals_total'] = gold + silver + bronze
        rows.append(row)
    except:
        print('-----')
        print('ERROR WITH', athlete)
        print('-----')    

len(rows)

df_medals = pd.DataFrame(rows)
df_medals.head(1)

df.head(1)

merged = df.merge(df_medals, left_on='endpoint_cleaned', right_on='endpoint')
merged.shape

# Check to see the total medal counts match up
merged[merged.medal_count != merged.medals_total]

# FACT CHECK

The medal count for Jernberg should be 9. (9 medals - 4 gold, 3 silver, 2 bronze)
* http://www.espn.com/olympics/winter/2010/athletes/_/athlete/33164/sixten-jernberg

The medal count for Thunberg should be 7. (7 medals - 5 gold, 1 silver, 1 bronze)
* http://www.espn.com/olympics/winter/2010/athletes/_/athlete/32565/clas-thunberg

The medal count for HAKULINEN should be 7. 3 Gold, 3 Silver, 1 Bronze (7 Total)
* https://www.sports-reference.com/olympics/athletes/ha/veikko-hakulinen-1.html

## Manually edit incorrect values 

rows[19]['medals_bronze'] = 2
rows[19]['medals_total'] = 9
rows[19]

rows[28]['medals_gold'] = 5
rows[28]['medals_silver'] = 1
rows[28]['medals_bronze'] = 1
rows[28]['medals_total'] = 7
rows[28]

rows[-4]['medals_silver'] = 3
rows[-4]['medals_total'] = 7
rows[-4]

# Re-do the dataframe transformation so that the correct medal counts are in the final, 'merged' dataframe

df_medals = pd.DataFrame(rows)
df_medals.shape

merged = df.merge(df_medals, left_on='endpoint_cleaned', right_on='endpoint')

merged.shape

# Check to see the total medal counts match up
merged[merged.medal_count != merged.medals_total]

## Now, medal_count and medals_total match.
* Remove medal_count, endpoint_x, endpoint_y columns to avoid confusion

merged = merged.drop(columns=["medal_count", 'endpoint_x', 'endpoint_y'])

merged['gender'] = merged.gender.str.title()
merged['game_type'] = 'Olympic'
merged['season'] = 'Winter'

merged

## Add the other_info and alternate_name columns so that the columns match Paralympic dataframes

merged['other_info'] = ''
merged['alternate_name'] = ''

merged.isna().sum()

merged.to_csv('oly_winter_complete.csv', index=False)

## Manually clean up the oly_winter_complete.csv 

* Add other_info and alternate_name for Victor An/Hyun-Soo Ahn
* Eliminate the duplicate rows for Victor An
* Save cleaned file as oly_winter_complete_cleaned.csv

df = pd.read_csv('oly_winter_complete_cleaned.csv')
df
