==> data_and_databases_project/jupyter_notebooks_funakoshi/Part 5 - Combine all df - final.ipynb <==
# Lede Summer 2019 Project - Part 5
## Combine all dataframes of Oly/Para info into one df

import requests
import pandas as pd
import re
import numpy as np
import os

import itertools

from bs4 import BeautifulSoup
from selenium import webdriver

df_para_summer = pd.read_csv('para_summer_complete.csv')
df_para_winter = pd.read_csv('para_winter_complete.csv')
df_oly_summer = pd.read_csv('oly_summer_complete.csv')
df_oly_winter = pd.read_csv('oly_winter_complete_cleaned.csv')

df_para_summer.head(3)

df_para_winter.head(1)

df_oly_summer.head(1)

df_oly_winter.head(1)

## Clean headers for Oly df so they match the Para ones

df_para_summer.columns

df_para_winter.columns

df_oly_summer.columns

df_oly_winter.columns

df_oly_summer = df_oly_summer.drop(columns = 'endpoint_cleaned')

df_oly_winter = df_oly_winter.drop(columns = 'endpoint_cleaned')

df_oly_summer.head(1)

df_oly_winter.head(1)

merged_para = df_para_summer.append(df_para_winter, sort=False)
# merged_para.dtypes

merged_oly = df_oly_winter.append(df_oly_summer, sort=False)
# merged_oly.dtypes

merged = merged_para.append(merged_oly, sort=True)

merged

## Clean up the dataframe by getting rid of whitespaces with strip()

merged['citizenship'] = merged.citizenship.str.strip()
merged['years'] = merged.years.str.strip()
merged['full_name'] = merged.full_name.str.strip()
merged['event'] = merged.event.str.strip()

split_years = merged.years.str.split('-')
for year in split_years:
    try:
        if year[1] < year[0]:
            print(year)
    except:
        pass

merged.to_csv('medal_info.csv', index=False)

## There is only ONE error where the century is wrong -- it should be 1992-2004.

* Export the dataframe as medal_info.csv
* Then manually edit in text editor
* Save cleaned file as medal_info_cleaned.csv
